Adding actions and actions test file.
